Python: instantiate future before predicting
The original code had
```python
forecast = m.fit(df).predict(future)
```
Well. Where did that `future` come from? A reader will be confused, because the output of that will depend of which variable `future` points to. If following the previous examples, this will point to a different dataset, leading to unexpected results.

This proposal makes the origin of the `future` explicit.